refactor: update import paths for types in CountrySelector and Footer components for better organization and clarity
